Fix grantlock 函数与定义不匹配

+++ src/syn.c
@@ -84,7 +84,7 @@ void sendwtnts(int operation);
 void popstack();
 void jia_unlock(int lock);
 void jia_barrier();
-void grantlock(long lock, int toproc, int acqscope);
+void grantlock(int lock, int toproc, int acqscope);
 void grantbarr(unsigned long lock);
 void acqserver(jia_msg_t *req);
 void relserver(jia_msg_t *req);
